c9

+++ src/rust/cctrusted_vm/src/tdvm.rs
@@ -255,11 +255,11 @@ impl CVM for TdxVM {
         // get quote wit vsock instead of TDVMCALL
         if !tdvmcall_flag {          
             const header_size: u32 = 4;
-            let qgs_msg_bytes_array: [u8; 16 + 8 + TDX_REPORT_LEN] = unsafe { transmute(qgs_msg) };
+            let qgs_msg_bytes_array: [u8; 16 + 8 + TDX_REPORT_LEN as usize] = unsafe { transmute(qgs_msg) };
             let msg_size = qgs_msg_bytes_array.len();
-            let msg_size_bytes_array: [u8; header_size] = unsafe { transmute(msg_size.to_be()) }; 
+            let msg_size_bytes_array: [u8; header_size as usize] = unsafe { transmute(msg_size.to_be()) }; 
 
-            let p_blob_payload = [0; header_size + 16 + 8 + TDX_REPORT_LEN];
+            let p_blob_payload = [0; header_size + 16 + 8 + TDX_REPORT_LEN as usize];
             p_blob_payload[..4].copy_from_slice(&msg_size_bytes_array);
             p_blob_payload[4..].copy_from_slice(&qgs_msg_bytes_array);
